Update to use the original model_path to decompose.
