+++ graph_net/test/subgraph_decompose_and_evaluation_step_test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-export PYTHONPATH=/root/paddlejob/workspace/env_run/liuyiqun/GraphNet:$PYTHONPATH
+export PYTHONPATH=/work/GraphNet:/work/abstract_pass/Athena:$PYTHONPATH
 
 GRAPH_NET_ROOT=$(python3 -c "import graph_net; import os; print(os.path.dirname(graph_net.__file__))")
 
@@ -52,7 +52,7 @@ test_target_device_config_str=$(cat <<EOF
 EOF
 )
 
-test_module_name="test_target_device"
+test_module_name="test_reference_device"
 if [ "${test_module_name}" = "test_compiler" ]; then
     TEST_CONFIG_B64=$(echo "$test_compiler_config_str" | base64 -w 0)
 elif [ "${test_module_name}" = "test_reference_device" ]; then
